requirements.txt
fastapi
uvicorn
python-telegram-bot
redis
sqlalchemy
psycopg2-binary
groq
python-dotenv
openai
requests
python-telegram-bot==20.0
